requirements.txt
slack_bolt
python-dotenv
openai